Update README.md
Suppose we want to give each student a number based on his/her personal data, such as weight and hight. How can we do that? One simple strategy is to nd a weighted sum,

s = 1w + 2h;

that combines both factors. The question is how to decide the ratio between 1 and 2?

One solution is to perform the Principle Component Analysis (PCA) of the given data. Suppose we have N students, whose weights and heights are (w1; h1), (w2; h2), : : :, (wN ; hN ). The PCA computes a vector (1-D subspace), so that the data projected to it have the maximum variance, as shown in Figure 1. The solution of the vector is called the principle component of the given data.


Figure 1: The principle components of the given data.


Now let’s translate this problem to the language of linear algebra. Let xi = (wi; hi)T and the principle component be y = ( 1; 2), where kyk = 1. Recall the projection of xi to y is (see Section 5.1)

P xi = (yyT )xi = (yT xi)y = aiy: